Moves shape-distance tests from ProximityEngine tests and into callback tests
For historical reasons, proximity_engine_test.cc contained many tests
that would exercise the underlying Callback mechanism. They don't
belong there. They belong with the unit tests for that Callback class.

The removes tests from proximity_engine_test.cc and places them in (a
previously missing) distance_to_shape_callback_test.cc and then
reframes them in terms of the callback. Some tests remain in the
ProximityEngine tests -- those tests that explicitly evaluate the
engine's responsibilities in invoking the Callback.
